+++ libs/langgraph/langgraph/graph/state.py
@@ -77,8 +77,8 @@ from langgraph.pregel.write import (
 )
 from langgraph.store.base import BaseStore
 from langgraph.types import All, Checkpointer, Command, RetryPolicy
-from langgraph.utils.fields import get_field_default
-from langgraph.utils.pydantic import create_model, get_update_as_tuples
+from langgraph.utils.fields import get_field_default, get_update_as_tuples
+from langgraph.utils.pydantic import create_model
 from langgraph.utils.runnable import RunnableLike, coerce_to_runnable
 
 logger = logging.getLogger(__name__)
