+++ libs/langgraph/langgraph/utils/pydantic.py
@@ -1,15 +1,11 @@
 import sys
 import typing
 from dataclasses import is_dataclass
-from typing import Any, Dict, Optional, Sequence, Union
+from typing import Any, Dict, Optional, Union
 
 import typing_extensions
 from pydantic import BaseModel
 from pydantic.v1 import BaseModel as BaseModelV1
-
-# NOTE: this is redefined here separately from langgraph.constants
-# to avoid a circular import
-MISSING = object()
 
 
 def create_model(
@@ -43,36 +39,6 @@ def create_model(
             v1_kwargs["__root__"] = root
 
         return create_model(model_name, **v1_kwargs, **(field_definitions or {}))
-
-
-def get_update_as_tuples(input: Any, keys: Sequence[str]) -> list[tuple[str, Any]]:
-    """Get Pydantic state update as a list of (key, value) tuples."""
-    # Pydantic v1
-    if isinstance(input, BaseModelV1):
-        keep: Optional[set[str]] = input.__fields_set__
-        defaults = {k: v.default for k, v in input.__fields__.items()}
-    # Pydantic v2
-    elif isinstance(input, BaseModel):
-        keep = input.model_fields_set
-        defaults = {k: v.default for k, v in input.model_fields.items()}
-    else:
-        keep = None
-        defaults = {}
-
-    # NOTE: This behavior for Pydantic is somewhat inelegant,
-    # but we keep around for backwards compatibility
-    # if input is a Pydantic model, only update values
-    # that are different from the default values or in the keep set
-    return [
-        (k, value)
-        for k in keys
-        if (value := getattr(input, k, MISSING)) is not MISSING
-        and (
-            value is not None
-            or defaults.get(k, MISSING) is not None
-            or (keep is not None and k in keep)
-        )
-    ]
 
 
 def is_supported_by_pydantic(type_: Any) -> bool:
